LibWeb: Share visible editing positions
Separate caret navigation from SelectionModifier so editing
operations can resolve rendered positions without mutating the live
selection. Preserve text affinity alongside every destination because
a DOM boundary alone cannot distinguish both sides of a soft line wrap.

Introduce VisiblePosition to retain both an exact DOM boundary and its
deep caret equivalent. Use the shared model when recording collapsed
selection snapshots so undo restores the canonical caret exposed after
editing.
